helper function for TOC link to detail page
