+++ docs/index.html
@@ -10,7 +10,7 @@
   <body>
     <div id="rsg-root"></div>
     <script src="build/bundle.b50e26a4.js"></script>
-    <script src="build/2.340d89ff.js"></script>
-    <script src="build/main.ec38cd46.js"></script>
+    <script src="build/2.edede095.js"></script>
+    <script src="build/main.fb7dbb4a.js"></script>
   </body>
 </html>
